Método de Pagamento

+++ src/main/java/br/unitins/topicos1/dto/VendaDTO.java
@@ -5,11 +5,11 @@ import java.util.List;
 import jakarta.validation.constraints.NotNull;
 
 public record VendaDTO(
-                @NotNull(message = "O id do tipo de pagamento não pode ser nulo")
-                 Integer idTipoPagamento,
+        @NotNull(message = "O id do tipo de pagamento não pode ser nulo")
+        Long idTipoPagamento,
 
-                List<ItemVendaDTO> itens,
-                
-                String cupom) {
+        List<ItemVendaDTO> itens,
+
+        String cupom) {
 
 }
